Move Value converter to PyProperty
PyValue class no longer necessary
